Moves DomainClassifier to Shoebox Module
